requirements.txt
Django==2.0
pytz==2017.3
gunicorn==19.7.1
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-oauthlib==0.2.0
google-auth-httplib2==0.0.3
requests==2.18.4
line-bot-sdk==1.5.0